+++ docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## v2.4.0
 * Removed web only mode for auth requests. Marking notifications as read is now disabled by default to deal with phishing accusations.
+* Save images with the correct extensions.
 
 ## v2.3.2
 * Disable auto feed refresh by default and add setting to re-enable it